Bug Fixes
Fixed the know bugs shown in the docs I have removed the issues from the docs.

To recap the issues were with certain static variable's that would stay true
for the powerup system.

THE BUGS: Getting the powerup and leaving and coming back to the main scene would make the player unkillable and would destroy obstacle's as if they were in powerup mode.

The other known bug was getting a game over would freeze other scenes when going to another scene after getting a gameover. Due to the Time.timeScale was = 0 I simply removed this from the gameover logic to fix it.

Changes in player controller script for powerup bug fix: Lines 50 thru 55

Changes in GameOverAnimation script for time froze bug fix: Removed  Lines 44: Time.timeScale = 0;