First MicroGame (WIP)
Work in progress math MircoMiniGame we have a functioning game, but at the moment it does not happen at the intended design. aka when we hit the MicroGame Test Obj.

Scripts Changed:
-Micro Game (Bug Fixes)
-MicroMath

GameObject we work with mostly is called MiniGameUI_Holder

*NOTE*
Found a bug - I managed to trigger the Cutscene we made with the MicroGame Prefab object while in PowerUp mode and I believe IEnumerator overriding is happening because of the fact that The PowerUp Mode messes with Speed as well in the same way the MircoGame script does.