added level ending animation(player model & camera animations)
Updated the afterimage effect somewhat and added a level completed UI as well as a new script that will be used as a reference for most of our UI elements that we reuse constantly.

New script is called UIElements and updated scripts were CutSceneScripts, EndOfLevel, and PlayerController and AfterImageEffect.

A new camera is also attached as a child to the MainCamera its name is GoalCamera used for when the level is completed animations.