Changed Animation for Dash/Jump and added Skybox TriggerScript
The change in animation for the dash and jump makes it so the dash and jump have 100% zero lag when transitioning to the animation kept the lag for the slide animation for the time being.

Added a Skybox Trigger Script and checks for whn the player collides with the hitbox and once the player does collide with it the skybox should change to the opposite time of day.